activity6/backend/requirements.txt
Flask==3.1.0
pymongo==3.11.4
PyJWT==2.9.0
flask-cors==6.0.0
Flask-Bcrypt==1.0.1